Update how grub displays menus
When we only have a single boot-environment, we won't display
the menu, unless the user holds down the SHIFT key. The system
will auto-boot after 2 seconds.

If multiple BE's are found, it will display the menu with a 5 second
timeout, allowing the user to halt and choose another BE.

Suggested by: Jordan Hubbard