Do some more Xorg voodoo for (NVIDIA/AMD)/Intel Hybrids.
Basically, during the first xorg setup, check if we have a second video
card, and if it is an "Intel" video. If so, we will attempt to start X using that
card first.

This should fix some issues we've been seeing with hybrids, where it "black-screens" the system
when it attempts to use the first device, which often cannot be accessed directly.