Add the ability for the mounttray to detect non-filesystem cd/dvd devices as audio/video disks, and prompt the user to open up an application to read that disk.
The application list is currently just the ~/bin/* contents. I need to write some XDG-application scanning functionality into libpcbsd before I can make this list of applications a lot smarter and more comprehensive.

It appears that the raw device cannot be passed to the application on the command-line (I.E. "vlc /dev/cd0") and have it work properly. These kind of disks need to be opened through special menu's in these applications (tested VLC and UMPlayer). So just have the mounttray startup the application without specifying the raw device as an input.