Bridge is destroyed when it still contains members
Noticed this issue on my FreeBSD system, running warden from sysutils/pcbsd-utils. The IP for my host is on bridge0 and I noticed that when I stopped the last warden jail my bridge0 would be deleted, but I still need it and it has my physical interface as a member so the member count is equal to 1 after the epair device is deleted.