Add support for "tempinit" and "temprem" options to personacrypt
These will do the following:

  tempinit <username> <size>

Will create a ZFS zvol of <size>, setup a GELI onetime key, format
the resulting file-system with UFS, and mount it on <username> $HOME

  temprem

Will remove the created temp dataset

This temporary file-system allows a user to login with a "clean" $HOME
directory. At logout, the file-system will be unmounted and destroyed.
Because of using a GELI onetime key, if the system reboots or restarts
without cleaning up, the old dataset on disk still cannot be read.