Update and rename default.css_t to classic.css_t
py-sphinx devs decided their old "default" theme is defunct and so produced a message to revise our conf.py to either continue to use the former default now named "classic" or switch to a new theme they consider default.  This sort of change definitely can be avoided in the future by creating our own custom theme so that we are no longer dependent upon the whims of the upstream devs with regard to themes.

This commit fixes a side-effect of the change by adjusting our theme-mod file to reflect the new name.