Update classic.css_t
It appears that there are other changes to py-sphinx theme generation so that the former CSS tags that were used have changed.  It is good that there are unique identifiers but not so good that there wasn't notice of the modification.

This adjustment returns the style (boxed and blue shaded) for our file names and keystrokes (now denoted by .kbd and .file), tt now given the former styling of pre at the least to make it visible. Very strangely, singlehtml renders (.kbd & .file) as desired while html does not?  Why?