Fix an unprocessed substitution
py-sphinx is unable tp expand "|version|" into 10.2 because the substitution is nested within other .rst markup.  Even though |version| may be properly interpreted by those reading the docs, this was not the intent.  In future we must note that this file will need instances of <version number> (presently *10.2*) replaced with <release version number> prior to publishing.