Fix bug #9977
This corrects a problem when installing to a specific GPT partition
and a SWAP partition was requested. Now the backend will shrink
the specified partition and then add a new swap partition in the
freed space. If the user sets the swapsize to 0, or disables it
in the GUI, then the original partition will not be re-sized.