remove remnant replaced by rst_epilog in conf.py
the .. include global.rst was intended as a way to define a bunch of values but primarily as a limited test  (long ago now).  global.rst does not exist in order for it to be included, but a number of values are now defined at the bottom/tail of conf.py in the rst_epilog section.