Fix bug #11372
Now when we install to a Partition/Slice/Free-Space, we will ignore
the particular method in which the install media was booted, and
instead use EFI or BIOS(CSM) mode depending upon the boot partition
on the disk. This fixes an issue where the disk is EFI formatted, but
the installer was booted in BIOS/CSM mode and it would try to setup
bios-boot and other non-EFI bits on disk.