Try to fix bug #11341
This will first get rid of the nasty kludge to dd the slice, and
instead create / destroy / create the BSD labed to destroy
any backup meta-data. (As freebsd does)

Next, for the first MBR/slice partition, throw a -b 8 on the
gpart add command, make sure we aren't chomping the BSD label
metadata.