Rename README.md to README
README.md needs to be renamed to README because it is not formatted for Markdown. When you view the file on GitHub a lot of formatting is lost because GitHub thinks the file is formatted for Markdown. This patch fixes README by making it be a regular file instead of a Markdown formatted file.